Refactor vue_safe module to serializers package. Refactor some utilies in lib package. Fix fooUrl to fooURL.
