+++ tests/fakes/providers.py
@@ -9,7 +9,7 @@ from undyingkingdoms.blueprints.admin.metadata import bot_county_prefix, bot_cou
 from undyingkingdoms.calculations.distributions import bell_curvish, normalize, curve_bounds
 from undyingkingdoms.metadata.metadata import metadata_races, metadata_titles, metadata_backgrounds
 from undyingkingdoms.metadata.research.metadata_research_all import generic_technology
-from utilities.helpers import romanize
+from lib.namers import romanize
 
 fake = faker.Faker()
