+++ undyingkingdoms/api/__init__.py
@@ -3,7 +3,7 @@ from importlib import import_module
 from flask import Blueprint
 
 from extensions import flask_csrf
-from utilities.helpers import to_class_name
+from lib.namers import to_class_name
 
 api_blueprint = Blueprint('api', __name__, url_prefix='/api')
 flask_csrf.exempt(api_blueprint)
