+++ undyingkingdoms/api/economy/gold.py
@@ -2,7 +2,7 @@ from flask import jsonify
 from flask.views import MethodView
 from flask_login import login_required, current_user
 
-from undyingkingdoms.api.vue_safe import vue_safe_metadata_mod, vue_safe_form
+from undyingkingdoms.serializers.vue_safe import vue_safe_metadata_mod, vue_safe_form
 from undyingkingdoms.models.forms.economy import EconomyForm
 from undyingkingdoms.metadata.metadata import income_modifier, tax_options
