+++ undyingkingdoms/api/economy/happiness.py
@@ -2,7 +2,7 @@ from flask import jsonify
 from flask.views import MethodView
 from flask_login import login_required, current_user
 
-from undyingkingdoms.api.vue_safe import vue_safe_metadata_mod
+from undyingkingdoms.serializers.vue_safe import vue_safe_metadata_mod
 from undyingkingdoms.metadata.metadata import happiness_modifier
