+++ undyingkingdoms/api/forum/threads.py
@@ -2,7 +2,7 @@ from flask import jsonify
 from flask.views import MethodView
 from flask_login import login_required
 
-from undyingkingdoms.api.vue_safe import vue_safe_thread
+from undyingkingdoms.serializers.vue_safe import vue_safe_thread
 from undyingkingdoms.models.forum import Thread
