+++ undyingkingdoms/api/helpers.py
@@ -1,6 +1,6 @@
 from importlib import import_module
 
-from utilities.helpers import to_class_name
+from lib.namers import to_class_name
 
 
 def import_endpoints(blueprint, mod_name, endpoints):
