+++ undyingkingdoms/serializers/vue_safe.py
@@ -1,7 +1,7 @@
 from flask import url_for
 from flask_login import current_user
 
-from utilities.helpers import to_mixed_case
+from lib.namers import to_mixed_case
 
 
 def vue_safe_nbsp(s):
@@ -92,17 +92,6 @@ def vue_safe_news(news):
             title=event.title,
             content=event.content
         )
-
-
-def vue_safe_message(message):
-    return dict(
-        time=message.time_created,
-        leader=message.get_county_leader_name(),
-        content=message.content,
-        room="global" if message.is_global else "kingdom",
-        id=message.id,
-        leaderUrl=url_for('enemy_overview', county_id=message.county_id)
-    )
 
 
 def vue_safe_reply(post):
